--- Assets/Sprint 3/CORE BUILD/LOCKED CORE ASSETS/Scripts/Player Scripts/Item Logic.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 88e7923e2a7e42f4b92f46a8d3c2bc09
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
