--- Assets/Sprint 3/CORE BUILD/LOCKED CORE ASSETS/Scripts/Player Scripts/Item Logic/Door.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 2f6b724c038fe46268b7c45727e05a6a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
