--- Assets/Sprint 3/CORE BUILD/LOCKED CORE ASSETS/Scripts/Player Scripts/Item Logic/HeavyItem.cs
@@ -1,20 +1,0 @@
-using UnityEngine;
-
-public class HeavyItem : MonoBehaviour
-{
-    // idk if yall want this, but I thought it was cool
-    [SerializeField, Tooltip("0 weight has no effect on movespeed, 100 weight slows movespeed down by 2x")]
-    public float weight;
-    
-    // Start is called before the first frame update
-    void Start()
-    {
-        
-    }
-
-    // Update is called once per frame
-    void Update()
-    {
-        
-    }
-}
