--- Assets/Sprint 3/CORE BUILD/LOCKED CORE ASSETS/Scripts/Player Scripts/Item Logic/HeavyItem.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 8ada4e404a84b49cfa51d0ef1160e59f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
