--- Assets/Sprint 3/CORE BUILD/LOCKED CORE ASSETS/Scripts/Player Scripts/Item Logic/Item SO.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: efa6f7493e7571642af63052a98839b3
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
