--- Assets/Sprint 3/CORE BUILD/LOCKED CORE ASSETS/Scripts/Player Scripts/Item Logic/Item SO/InventoryBase.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 062073fe972044885a8d4efed7e0c829
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
