--- Assets/Sprint 3/CORE BUILD/LOCKED CORE ASSETS/Scripts/Player Scripts/Item Logic/ItemPickup.cs
@@ -1,9 +1,0 @@
-using UnityEngine;
-
-namespace Sprint_3.CORE_BUILD.LOCKED_CORE_ASSETS.Scripts.Player_Scripts.Item_Logic
-{
-    public class ItemPickup : MonoBehaviour
-    {
-        public ItemBase item;
-    }
-}
