--- Assets/Sprint 3/CORE BUILD/LOCKED CORE ASSETS/Scripts/Player Scripts/Item Logic/ItemPickup.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e641391fcb5cb473b92291a9aebd5aef
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
