--- Assets/Sprint 3/CORE BUILD/LOCKED CORE ASSETS/Scripts/Player Scripts/Item Logic/New Consumable.asset
@@ -1,23 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 0}
-  m_Name: New Consumable
-  m_EditorClassIdentifier: Assembly-CSharp::Consumable
-  ItemID: fleshdust
-  rarity: 0
-  description: 
-  stackSize: 0
-  thumbnail: {fileID: 0}
-  prefab: {fileID: 0}
-  type: 0
-  attractionStrength: 0
-  soundLevel: 0
