--- Assets/Sprint 3/CORE BUILD/LOCKED CORE ASSETS/Scripts/Player Scripts/Item Logic/New Consumable.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: a6267e87ac339447a878c80060349000
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
