+++ Assets/Sprint 3/Inventory.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f67d800789c954cb98f498e4327a0b36
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
