+++ Assets/Sprint 3/Inventory/Item Data.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8327a976e2b6a435886c2a498aeffaf1
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
