+++ Assets/Sprint 3/Inventory/Item Data/3TestItem.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6a52328f137b249269b2e585a37cf5c1
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
