+++ Assets/Sprint 3/Inventory/Item Data/Item2Test.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 02c94fe59e3f4458ab7405abe74bcbff
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
