+++ Assets/Sprint 3/Inventory/Item Data/NOTItemTest.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: dbe278cc1d60f4780a4d186ba52c7e20
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
