+++ Assets/Sprint 3/Inventory/Item Data/TestItem.asset
@@ -1,0 +1,24 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 475b153b495cd4d1996f2b7a89b79eb0, type: 3}
+  m_Name: TestItem
+  m_EditorClassIdentifier: 
+  ItemID: 
+  rarity: 0
+  cost: 0
+  description: 
+  stackSize: 0
+  thumbnail: {fileID: 0}
+  prefab: {fileID: 0}
+  type: 0
+  attractionStrength: 0
+  soundLevel: 0
