+++ Assets/Sprint 3/Inventory/Item Data/TestItem.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 44acde0a12bf0492f857557ef6d08969
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
