+++ Assets/Sprint 3/Inventory/Scripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 72d8c97317f2f46a6a05899372e83b62
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
