+++ Assets/Sprint 3/Inventory/Scripts/Item SO.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: efa6f7493e7571642af63052a98839b3
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
