+++ Assets/Sprint 3/Inventory/Scripts/Item SO/Artifact.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e5e4a30813d594bba8e0f309a88d0edc
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
