+++ Assets/Sprint 3/Inventory/Scripts/Item SO/Consumable.cs
@@ -1,0 +1,10 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+// CONSUMABLES CLASS
+[CreateAssetMenu(fileName = "New Consumable", menuName = "Inventory/Items/Consumable")]
+public class Consumable : ItemBase
+{
+    
+}
