+++ Assets/Sprint 3/Inventory/Scripts/Item SO/Consumable.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 475b153b495cd4d1996f2b7a89b79eb0
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
