+++ Assets/Sprint 3/Inventory/Scripts/Item SO/ItemBase.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ef8904b46c1bf4eb78fd717b00fabe54
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
