+++ Assets/Sprint 3/Inventory/Scripts/ItemObject.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 36a58cfdf2c274e6ba944b15c37319c6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
