+++ Assets/Sprint 3/Inventory/Scripts/ItemPickup.cs
@@ -1,0 +1,9 @@
+using UnityEngine;
+
+namespace Sprint_3.CORE_BUILD.LOCKED_CORE_ASSETS.Scripts.Player_Scripts.Item_Logic
+{
+    public class ItemPickup : MonoBehaviour
+    {
+        public ItemBase item;
+    }
+}
