+++ Assets/Sprint 3/Inventory/Scripts/ItemPickup.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e641391fcb5cb473b92291a9aebd5aef
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
